requirements.txt
decorator==4.3.0
Django==2.1.5
django-filter==2.0.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
psycopg2-binary==2.7.6.1
public==2018.11.20
PyJWT==1.7.1
pytz==2018.9
self==2018.11.20
